+++ server/src/main/java/org/apache/iotdb/db/mpp/execution/operator/source/AlignedSeriesAggregationScanOperator.java
@@ -26,7 +26,6 @@ import org.apache.iotdb.db.mpp.plan.planner.plan.node.PlanNodeId;
 import org.apache.iotdb.db.mpp.plan.planner.plan.parameter.GroupByTimeParameter;
 import org.apache.iotdb.tsfile.file.metadata.statistics.Statistics;
 import org.apache.iotdb.tsfile.read.common.block.TsBlock;
-import org.apache.iotdb.tsfile.read.common.block.TsBlock.TsBlockSingleColumnIterator;
 import org.apache.iotdb.tsfile.read.filter.basic.Filter;
 
 import java.io.IOException;
@@ -36,8 +35,7 @@ import java.util.List;
 import static org.apache.iotdb.db.mpp.execution.operator.AggregationUtil.isAllAggregatorsHasFinalResult;
 
 /** This operator is responsible to do the aggregation calculation especially for aligned series. */
-public class AlignedSeriesAggregationScanOperator extends SeriesAggregationScanOperator
-    implements DataSourceOperator {
+public class AlignedSeriesAggregationScanOperator extends AbstractSeriesAggregationScanOperator {
 
   private final int subSensorSize;
 
@@ -62,69 +60,7 @@ public class AlignedSeriesAggregationScanOperator extends SeriesAggregationScanO
   }
 
   @Override
-  protected void calculateNextAggregationResult() {
-    try {
-      if (calcFromCacheData()) {
-        updateResultTsBlock();
-        return;
-      }
-
-      // read page data firstly
-      if (readAndCalcFromPage()) {
-        updateResultTsBlock();
-        return;
-      }
-
-      // read chunk data secondly
-      if (readAndCalcFromChunk()) {
-        updateResultTsBlock();
-        return;
-      }
-
-      // read from file first
-      while (seriesScanUtil.hasNextFile()) {
-        if (canUseCurrentFileStatistics()) {
-          Statistics fileTimeStatistics = seriesScanUtil.currentFileTimeStatistics();
-          if (fileTimeStatistics.getStartTime() > curTimeRange.getMax()) {
-            if (ascending) {
-              updateResultTsBlock();
-              return;
-            } else {
-              seriesScanUtil.skipCurrentFile();
-              continue;
-            }
-          }
-          // calc from fileMetaData
-          if (curTimeRange.contains(
-              fileTimeStatistics.getStartTime(), fileTimeStatistics.getEndTime())) {
-            Statistics[] statisticsList = new Statistics[subSensorSize];
-            for (int i = 0; i < subSensorSize; i++) {
-              statisticsList[i] = seriesScanUtil.currentFileStatistics(i);
-            }
-            calcFromStatisticsArray(statisticsList);
-            seriesScanUtil.skipCurrentFile();
-            if (isAllAggregatorsHasFinalResult(aggregators) && !isGroupByQuery) {
-              break;
-            } else {
-              continue;
-            }
-          }
-        }
-
-        // read chunk
-        if (readAndCalcFromChunk()) {
-          updateResultTsBlock();
-          return;
-        }
-      }
-
-      updateResultTsBlock();
-    } catch (IOException e) {
-      throw new RuntimeException("Error while scanning the file", e);
-    }
-  }
-
-  private void calcFromStatisticsArray(Statistics[] statistics) {
+  protected void calcFromStatistics(Statistics[] statistics) {
     for (Aggregator aggregator : aggregators) {
       if (aggregator.hasFinalResult()) {
         continue;
@@ -133,28 +69,28 @@ public class AlignedSeriesAggregationScanOperator extends SeriesAggregationScanO
     }
   }
 
-  private boolean readAndCalcFromPage() throws IOException {
-    while (seriesScanUtil.hasNextPage()) {
-      if (canUseCurrentPageStatistics()) {
-        Statistics pageTimeStatistics = seriesScanUtil.currentPageTimeStatistics();
-        // There is no more eligible points in current time range
-        if (pageTimeStatistics.getStartTime() > curTimeRange.getMax()) {
+  @Override
+  protected boolean readAndCalcFromFile() throws IOException {
+    while (seriesScanUtil.hasNextFile()) {
+      if (canUseCurrentFileStatistics()) {
+        Statistics fileTimeStatistics = seriesScanUtil.currentFileTimeStatistics();
+        if (fileTimeStatistics.getStartTime() > curTimeRange.getMax()) {
           if (ascending) {
             return true;
           } else {
-            seriesScanUtil.skipCurrentPage();
+            seriesScanUtil.skipCurrentFile();
             continue;
           }
         }
-        // can use pageHeader
+        // calc from fileMetaData
         if (curTimeRange.contains(
-            pageTimeStatistics.getStartTime(), pageTimeStatistics.getEndTime())) {
+            fileTimeStatistics.getStartTime(), fileTimeStatistics.getEndTime())) {
           Statistics[] statisticsList = new Statistics[subSensorSize];
           for (int i = 0; i < subSensorSize; i++) {
-            statisticsList[i] = seriesScanUtil.currentPageStatistics(i);
+            statisticsList[i] = seriesScanUtil.currentFileStatistics(i);
           }
-          calcFromStatisticsArray(statisticsList);
-          seriesScanUtil.skipCurrentPage();
+          calcFromStatistics(statisticsList);
+          seriesScanUtil.skipCurrentFile();
           if (isAllAggregatorsHasFinalResult(aggregators) && !isGroupByQuery) {
             return true;
           } else {
@@ -163,35 +99,17 @@ public class AlignedSeriesAggregationScanOperator extends SeriesAggregationScanO
         }
       }
 
-      // calc from page data
-      TsBlock tsBlock = seriesScanUtil.nextPage();
-      TsBlockSingleColumnIterator tsBlockIterator = tsBlock.getTsBlockSingleColumnIterator();
-      if (tsBlockIterator == null || !tsBlockIterator.hasNext()) {
-        continue;
-      }
-
-      // stop calc and cached current batchData
-      if (ascending && tsBlockIterator.currentTime() > curTimeRange.getMax()) {
-        cachedData = tsBlock;
-        return true;
-      }
-
-      // calc from batch data
-      calcFromBatch(tsBlock);
-
-      // judge whether the calculation finished
-      boolean isTsBlockOutOfBound =
-          ascending
-              ? tsBlock.getEndTime() > curTimeRange.getMax()
-              : tsBlock.getEndTime() < curTimeRange.getMin();
-      if (isAllAggregatorsHasFinalResult(aggregators) || isTsBlockOutOfBound) {
+      // read chunk
+      if (readAndCalcFromChunk()) {
         return true;
       }
     }
+
     return false;
   }
 
-  private boolean readAndCalcFromChunk() throws IOException {
+  @Override
+  protected boolean readAndCalcFromChunk() throws IOException {
     while (seriesScanUtil.hasNextChunk()) {
       if (canUseCurrentChunkStatistics()) {
         Statistics chunkTimeStatistics = seriesScanUtil.currentChunkTimeStatistics();
@@ -211,7 +129,7 @@ public class AlignedSeriesAggregationScanOperator extends SeriesAggregationScanO
           for (int i = 0; i < subSensorSize; i++) {
             statisticsList[i] = seriesScanUtil.currentChunkStatistics(i);
           }
-          calcFromStatisticsArray(statisticsList);
+          calcFromStatistics(statisticsList);
           seriesScanUtil.skipCurrentChunk();
           if (isAllAggregatorsHasFinalResult(aggregators) && !isGroupByQuery) {
             return true;
@@ -229,21 +147,69 @@ public class AlignedSeriesAggregationScanOperator extends SeriesAggregationScanO
     return false;
   }
 
-  private boolean canUseCurrentFileStatistics() throws IOException {
+  @Override
+  protected boolean readAndCalcFromPage() throws IOException {
+    while (seriesScanUtil.hasNextPage()) {
+      if (canUseCurrentPageStatistics()) {
+        Statistics pageTimeStatistics = seriesScanUtil.currentPageTimeStatistics();
+        // There is no more eligible points in current time range
+        if (pageTimeStatistics.getStartTime() > curTimeRange.getMax()) {
+          if (ascending) {
+            return true;
+          } else {
+            seriesScanUtil.skipCurrentPage();
+            continue;
+          }
+        }
+        // can use pageHeader
+        if (curTimeRange.contains(
+            pageTimeStatistics.getStartTime(), pageTimeStatistics.getEndTime())) {
+          Statistics[] statisticsList = new Statistics[subSensorSize];
+          for (int i = 0; i < subSensorSize; i++) {
+            statisticsList[i] = seriesScanUtil.currentPageStatistics(i);
+          }
+          calcFromStatistics(statisticsList);
+          seriesScanUtil.skipCurrentPage();
+          if (isAllAggregatorsHasFinalResult(aggregators) && !isGroupByQuery) {
+            return true;
+          } else {
+            continue;
+          }
+        }
+      }
+
+      // calc from page data
+      TsBlock tsBlock = seriesScanUtil.nextPage();
+      if (tsBlock == null || tsBlock.isEmpty()) {
+        continue;
+      }
+
+      // calc from raw data
+      if (calcFromRawData(tsBlock)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  @Override
+  protected boolean canUseCurrentFileStatistics() throws IOException {
     Statistics fileStatistics = seriesScanUtil.currentFileTimeStatistics();
     return !seriesScanUtil.isFileOverlapped()
         && fileStatistics.containedByTimeFilter(seriesScanUtil.getTimeFilter())
         && !seriesScanUtil.currentFileModified();
   }
 
-  private boolean canUseCurrentChunkStatistics() throws IOException {
+  @Override
+  protected boolean canUseCurrentChunkStatistics() throws IOException {
     Statistics chunkStatistics = seriesScanUtil.currentChunkTimeStatistics();
     return !seriesScanUtil.isChunkOverlapped()
         && chunkStatistics.containedByTimeFilter(seriesScanUtil.getTimeFilter())
         && !seriesScanUtil.currentChunkModified();
   }
 
-  private boolean canUseCurrentPageStatistics() throws IOException {
+  @Override
+  protected boolean canUseCurrentPageStatistics() throws IOException {
     Statistics currentPageStatistics = seriesScanUtil.currentPageTimeStatistics();
     if (currentPageStatistics == null) {
       return false;
